requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
sqlalchemy>=2.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0
cmdstanpy>=1.2.0
streamlit>=1.29.0
tensorflow>=2.14.0
torch>=2.1.0
notebook>=7.0.0
joblib>=1.3.0
